Update CVE-2021-21985.nse
Hi Alex,

Thanks for this POC. However, during my verification, I have identified that the script is not giving correct results. In some cases, the script is resulting the hosts as `Likely Vulnerable` even if they are not (when I checked manually). So I had a look at the script and noticed that one of the `IF` clause is not working as expected while checking variable "VULNERABLE2" against the response body, we should also add a clause to check response status as 200 along with this. I have added the same, and proposing my changes here. Please validate the same.

Thanks,
Shraddha